Rename prepared_statement.h to prepare.h
Use the same pattern as prepare.c.
